Conflicts:
	GPS_LCD/GPS_LCD.pde
	GPS_LCD/helpers.h

and THATS WHY you ALWAYS UPDATE FROM MASTER

+++ GPS_LCD/helpers.h
@@ -32,8 +32,8 @@ SIGNAL(WDT_vect) {
 char* formattedTime(char *time) {
 	String timeString = "";
 	char *timeCharArray = "00:00";
-	uint8_t hour = 0;
-	uint8_t minute = 0;
+	int hour = 0;
+	int minute = 0;
 	hour = hour + (10 * (time[0]-'0'));
 	hour = hour + (time[1]-'0');
 	hour = hour + TIMEZONE;
